fix: standardized load balancer dashboard widgets
