+++ src/dashboards/OperationAvailabilityAndLatencyDashboard.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { Fn } from 'aws-cdk-lib';
+import { Aws, Fn } from 'aws-cdk-lib';
 import {
   Dashboard,
   IMetric,
@@ -13,11 +13,6 @@ import {
   AlarmWidget,
   LegendPosition,
 } from 'aws-cdk-lib/aws-cloudwatch';
-import {
-  BaseLoadBalancer,
-  CfnLoadBalancer,
-  IApplicationLoadBalancer,
-} from 'aws-cdk-lib/aws-elasticloadbalancingv2';
 import { Construct } from 'constructs';
 import { ContributorInsightsWidget } from './ContributorInsightsWidget';
 import { IOperationAvailabilityAndLatencyDashboard } from './IOperationAvailabilityAndLatencyDashboard';
@@ -25,8 +20,6 @@ import { OperationAvailabilityAndLatencyDashboardProps } from './props/Operation
 import { OperationAvailabilityWidgetProps } from './props/OperationAvailabilityWidgetProps';
 import { OperationLatencyWidgetProps } from './props/OperationLatencyWidgetProps';
 import { IAvailabilityZoneMapper } from '../azmapper/IAvailabilityZoneMapper';
-import { ApplicationLoadBalancerMetrics } from '../metrics/ApplicationLoadBalancerMetrics';
-import { NetworkLoadBalancerMetrics } from '../metrics/NetworkLoadBalancerMetrics';
 import { RegionalAvailabilityMetrics } from '../metrics/RegionalAvailabilityMetrics';
 import { RegionalLatencyMetrics } from '../metrics/RegionalLatencyMetrics';
 import { ZonalAvailabilityMetrics } from '../metrics/ZonalAvailabilityMetrics';
@@ -187,7 +180,7 @@ export class OperationAvailabilityAndLatencyDashboard
         height: 6,
         width: 24,
         title: Fn.sub('${AWS::Region} TPS'),
-        region: Fn.sub('${AWS::Region}'),
+        region: Aws.REGION,
         left: [
           RegionalAvailabilityMetrics.createRegionalAvailabilityMetric({
             label: Fn.ref('AWS::Region'),
@@ -211,7 +204,7 @@ export class OperationAvailabilityAndLatencyDashboard
           height: 6,
           width: 8,
           title: availabilityZoneId + ' TPS',
-          region: Fn.sub('${AWS::Region}'),
+          region: Aws.REGION,
           left: [
             ZonalAvailabilityMetrics.createZonalAvailabilityMetric({
               availabilityZoneId: availabilityZoneId,
@@ -252,7 +245,7 @@ export class OperationAvailabilityAndLatencyDashboard
         height: 8,
         width: 24,
         title: Fn.sub('${AWS::Region} Availability'),
-        region: Fn.sub('${AWS::Region}'),
+        region: Aws.REGION,
         left: [
           RegionalAvailabilityMetrics.createRegionalAvailabilityMetric({
             label: Fn.ref('AWS::Region'),
@@ -295,7 +288,7 @@ export class OperationAvailabilityAndLatencyDashboard
       new AlarmWidget({
         height: 2,
         width: 24,
-        region: Fn.sub('${AWS::Region}'),
+        region: Aws.REGION,
         alarm: props.regionalEndpointAvailabilityAlarm,
         title: 'Success Rate',
       }),
@@ -313,7 +306,7 @@ export class OperationAvailabilityAndLatencyDashboard
           height: 6,
           width: 8,
           title: availabilityZoneId + ' Availability',
-          region: Fn.sub('${AWS::Region}'),
+          region: Aws.REGION,
           left: [
             ZonalAvailabilityMetrics.createZonalAvailabilityMetric({
               availabilityZoneId: availabilityZoneId,
@@ -363,7 +356,7 @@ export class OperationAvailabilityAndLatencyDashboard
             new AlarmWidget({
               height: 2,
               width: 8,
-              region: Fn.sub('${AWS::Region}'),
+              region: Aws.REGION,
               alarm: props.zonalEndpointAvailabilityAlarms[k],
               title: 'Success Rate',
             }),
@@ -451,7 +444,7 @@ export class OperationAvailabilityAndLatencyDashboard
           height: 8,
           width: 24,
           title: Fn.sub('${AWS::Region} Latency'),
-          region: Fn.sub('${AWS::Region}'),
+          region: Aws.REGION,
           left: latencyMetrics,
           leftYAxis: {
             max: props.latencyMetricDetails.successAlarmThreshold * 1.5,
@@ -475,7 +468,7 @@ export class OperationAvailabilityAndLatencyDashboard
       new AlarmWidget({
         height: 2,
         width: 24,
-        region: Fn.sub('${AWS::Region}'),
+        region: Aws.REGION,
         alarm: props.regionalEndpointLatencyAlarm,
       }),
     );
@@ -529,7 +522,7 @@ export class OperationAvailabilityAndLatencyDashboard
             height: 6,
             width: 8,
             title: availabilityZoneId + ' Latency',
-            region: Fn.sub('${AWS::Region}'),
+            region: Aws.REGION,
             left: latencyMetrics2,
             leftAnnotations: [
               {
@@ -558,7 +551,7 @@ export class OperationAvailabilityAndLatencyDashboard
             new AlarmWidget({
               height: 2,
               width: 8,
-              region: Fn.sub('${AWS::Region}'),
+              region: Aws.REGION,
               alarm: props.zonalEndpointLatencyAlarms[k],
             }),
           );
@@ -588,125 +581,6 @@ export class OperationAvailabilityAndLatencyDashboard
     }
 
     return latencyWidgets;
-  }
-
-  private static createApplicationLoadBalancerWidgets(
-    props: OperationAvailabilityAndLatencyDashboardProps,
-    title: string,
-    availabilityZoneNames: string[],
-  ): IWidget[] {
-    let albWidgets: IWidget[] = [];
-
-    albWidgets.push(new TextWidget({ height: 2, width: 24, markdown: title }));
-
-    albWidgets.push(
-      new GraphWidget({
-        height: 8,
-        width: 24,
-        title: Fn.sub('${AWS::Region} Processed Bytes'),
-        region: Fn.sub('${AWS::Region}'),
-        left: [
-          ApplicationLoadBalancerMetrics.getRegionalProcessedBytesMetric(
-            props.loadBalancer as IApplicationLoadBalancer,
-            props.operation.serverSideAvailabilityMetricDetails.period,
-          ),
-        ],
-        leftYAxis: {
-          label: 'Processed Bytes',
-          showUnits: true,
-        },
-      }),
-    );
-
-    availabilityZoneNames.forEach((availabilityZoneName) => {
-      let availabilityZoneId: string =
-        props.azMapper.availabilityZoneIdFromAvailabilityZoneLetter(
-          availabilityZoneName.substring(availabilityZoneName.length - 1),
-        );
-
-      albWidgets.push(
-        new GraphWidget({
-          height: 6,
-          width: 8,
-          title: availabilityZoneId + ' Processed Bytes',
-          region: Fn.sub('${AWS::Region}'),
-          left: [
-            ApplicationLoadBalancerMetrics.getPerAZProcessedBytesMetric(
-              props.loadBalancer as IApplicationLoadBalancer,
-              availabilityZoneName,
-              availabilityZoneId,
-              props.operation.serverSideAvailabilityMetricDetails.period,
-            ),
-          ],
-          leftYAxis: {
-            label: 'Processed Bytes',
-            showUnits: true,
-          },
-        }),
-      );
-    });
-
-    return albWidgets;
-  }
-
-  private static createNetworkLoadBalancerWidgets(
-    props: OperationAvailabilityAndLatencyDashboardProps,
-    title: string,
-    availabilityZoneNames: string[],
-  ): IWidget[] {
-    let nlbWidgets: IWidget[] = [];
-    let loadBalancerFullName: string = (props.loadBalancer as BaseLoadBalancer)
-      .loadBalancerFullName;
-
-    nlbWidgets.push(new TextWidget({ height: 2, width: 24, markdown: title }));
-
-    nlbWidgets.push(
-      new GraphWidget({
-        height: 8,
-        width: 24,
-        title: Fn.sub('${AWS::Region} Processed Bytes'),
-        region: Fn.sub('${AWS::Region}'),
-        left: [
-          NetworkLoadBalancerMetrics.createRegionalNetworkLoadBalancerProcessedBytesMetric(
-            loadBalancerFullName,
-            props.operation.serverSideAvailabilityMetricDetails.period,
-          ),
-        ],
-        leftYAxis: {
-          label: 'Processed Bytes',
-          showUnits: true,
-        },
-      }),
-    );
-
-    availabilityZoneNames.forEach((availabilityZoneName) => {
-      let availabilityZoneId: string =
-        props.azMapper.availabilityZoneIdFromAvailabilityZoneLetter(
-          availabilityZoneName.substring(availabilityZoneName.length - 1),
-        );
-
-      nlbWidgets.push(
-        new GraphWidget({
-          height: 6,
-          width: 8,
-          title: availabilityZoneId + ' Processed Bytes',
-          region: Fn.sub('${AWS::Region}'),
-          left: [
-            NetworkLoadBalancerMetrics.createZonalNetworkLoadBalancerProcessedBytesMetric(
-              loadBalancerFullName,
-              availabilityZoneName,
-              props.operation.serverSideAvailabilityMetricDetails.period,
-            ),
-          ],
-          leftYAxis: {
-            label: 'Processed Bytes',
-            showUnits: true,
-          },
-        }),
-      );
-    });
-
-    return nlbWidgets;
   }
 
   /**
@@ -776,29 +650,6 @@ export class OperationAvailabilityAndLatencyDashboard
         '**Server-side Latency**',
       ),
     );
-
-    let lb: CfnLoadBalancer = props.loadBalancer?.node
-      .defaultChild as CfnLoadBalancer;
-
-    if (lb !== undefined && lb != null) {
-      if (lb.type == 'application') {
-        widgets.push(
-          OperationAvailabilityAndLatencyDashboard.createApplicationLoadBalancerWidgets(
-            props,
-            '**Application Load Balancer Metrics**',
-            props.operation.service.availabilityZoneNames,
-          ),
-        );
-      } else if (lb.type == 'network') {
-        widgets.push(
-          OperationAvailabilityAndLatencyDashboard.createNetworkLoadBalancerWidgets(
-            props,
-            '**Network Load Balancer Metrics**',
-            props.operation.service.availabilityZoneNames,
-          ),
-        );
-      }
-    }
 
     if (
       props.operation.canaryMetricDetails !== undefined &&
